+++ Python/.dockerignore
@@ -1,2 +1,3 @@
 Dockerfile
 .dockerignore
+env
