--- Python/rvc/.env
@@ -1,8 +1,0 @@
-OPENBLAS_NUM_THREADS = 1
-no_proxy = localhost, 127.0.0.1, ::1
-
-# You can change the location of the model, etc. by changing here
-weight_root = assets/weights
-weight_uvr5_root = assets/uvr5_weights
-index_root = logs
-rmvpe_root = assets/rmvpe
