--- Python/rvc/assets/hubert/.gitignore
@@ -1,3 +1,0 @@
-*
-!.gitignore
-!hubert_inputs.pth
