--- Python/rvc/assets/rmvpe/.gitignore
@@ -1,3 +1,0 @@
-*
-!.gitignore
-!rmvpe_inputs.pth
