+++ Python/rvc_service/.env
@@ -1,0 +1,8 @@
+OPENBLAS_NUM_THREADS = 1
+no_proxy = localhost, 127.0.0.1, ::1
+
+# You can change the location of the model, etc. by changing here
+weight_root = 
+weight_uvr5_root = rvc_service/assets/uvr5_weights
+index_root = 
+rmvpe_root = rvc_service/assets/rmvpe
