+++ Python/rvc_service/assets/hubert/.gitignore
@@ -1,0 +1,3 @@
+*
+!.gitignore
+!hubert_inputs.pth
