+++ Python/rvc_service/assets/rmvpe/.gitignore
@@ -1,0 +1,3 @@
+*
+!.gitignore
+!rmvpe_inputs.pth
